+++ ll/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for `embedded-hal` version `1.0.0-alpha.9` with the `eh1` feature.
 - Added support for `embedded-hal-async` version `0.1.0-alpha.3` with the `eha0a` feature.
 - Added an `aio` module with async traits.
+- Added `sn_tx_ptrs` and `sn_rx_ptrs` to read socket buffer pointers in a single transaction.
 - Added `From<Sn>` for all primitive signed and unsigned integers.
 - Added `TryFrom<_> for Sn` for all primitive signed and unsigned integers.